test: merge image-insert-caret tests into image.test.ts
